Add full proxy support
This PR gets the Kodi global proxy settings and then applies those to
all invocations of requests.

This is the first implementation that adds full proxy support.

In the future I would like to add 2 additional features so that users
can select between:

- Use Kodi proxy settings (default)
- Disable proxy support for this addon
- Custom proxy settings (for this addon)

The current PR implements the first (default) option.