Add 'live' to live TV channels
Ensure 'livd' is in the menu entry name so that when these entries are
added to favourites, they clearly indicate this brings you to the live
channels.

At some point we will add listings per channel, so the distinction is
worthwhile in the future (the sooner the better for new users).