Replace Mock with xbmc library replacement (#259)
This PR includes:
- Implementation of xbmc libraries
- New unit tests

For testing the caching-functionality the use of Mock() became an even
more complex issue. So this replaces a quite complex Mock() setup to
mock kodiwrapper with an alternative implementation of the xbmc
libraries.

The benefit is that we don't have to fake every possible kodiwrapper
functionality, but only the xbmc functionality that kodiwrapper is
using. This means we are now also testing kodiwrapper to a greater
extent.