Add default values for missing settings
The xbmcaddon getSetting() function returns an empty string for
non-existing settings. Which makes it unclear if a setting is missing,
or simply set to an empty string on purpose.

But for boolean values, this is not a real issue.

This implements providing a default in case the setting does not exist,
or is empty.

This PR includes:
- Default values for get_settings()
- Improve kodi xbmcaddon stub
- Fix pylint issues related to long lines