Add a Refresh context menu
The original implementation was very invasive and touched almost every
API to pass a parameter `use_cache` so the specific cache would be
invalidated.

This is easier but as effective, we have a small list of actions with
known caches, and we simply reuse the existing params to refresh the
specific cache and then refresh the container.