Do not clear username if login failed
So currently when you type in a username and password, you can visually
verify the username is correct, but the password is a bunch of
asterisks. So if the login failed we assume the username was correct
(and can be verified by the user) but since the password is not visible
and cannot be edited we reset only the password.

This is a bit more userfriendly. An option to show the password would be
even better so we don't need to reset anything (just mark as invalid).