Move cursor down when ListItem is unfollowed/removed
This fixes an annoying issue where your cursor would jump to the start of
the list when you unfollow a program in the favorite listing, and so the
selected ListItem disappears from under the cursor.

This also implements a smarter way to know where we are within the
add-on.

And a potential small fix to better codecov reporting.