Add availability and Geoblocked on the same line
If it fits, this adds Geo-blocked on the same line as availability.

This works in English and not in Dutch (based on screen-size).
Since the colours give a visual separation, we don't need to add a
separator (which would ruin this).

I much rather added a Lock picture/character/emoji, but that is not
possible at this time.

This PR also added/changed a few comments for localize().