Add HTTP error to dialog
I got another HTTP Error 415: Unsupported Media Type, although we do not
report or log this error anywhere.

This change adds it to the dialog and logs it to the log. So that if it
gets reported (e.g. to VRT NU), hopefully the details will help.

This PR includes:
- Put the HTTP Error in the dialog
- Log the HTTP Error to a log file
- Don't handle resumepoints if stream is None